Changed support from SparseOctree<double, uint16_t> to SparseOctree<double, uint8_t>; Redefined ply export for SparseOctree<double, uint8_t>

+++ OctreeApi/OctreeLibrary/include/types.cuh
@@ -5,7 +5,7 @@
 
 struct Averaging
 {
-    float r, g, b;
+    uint64_t r, g, b;
     uint32_t pointCount;
 };
